updated loader
