+++ pages/Appointment/index.vue
@@ -1,7 +1,9 @@
 <script setup lang="ts">
 import HelloWorld from '~/core/Appointment/components/HelloWorld.vue';
+import UsersExample from '~/shared/UsersExample.vue';
 </script>
 
 <template>
     <HelloWorld></HelloWorld>
+    <UsersExample/>
 </template>
